Extracted protocols to its own file. Added interactor test
