+++ MovieList/MovieListTests/UpcomingMovies/MockData.swift
@@ -1,0 +1,16 @@
+//
+//MockData.swift
+//MovieListTests
+//
+//Created by Marcelo on 04/08/19.
+//Copyright © 2019 Marcelo. All rights reserved.
+//
+
+import Foundation
+
+let kUpcomingMovies: [MovieModel] = [
+    MovieModel(id: 1, posterPath: "", backdropPath: "", overview: "Overview 1", releaseDate: "2019-06-01", title: "Movie 1", genres: ["genre 1"]),
+    MovieModel(id: 2, posterPath: "", backdropPath: "", overview: "Overview 2", releaseDate: "2019-06-02", title: "Movie 2", genres: ["genre 2"]),
+    MovieModel(id: 3, posterPath: "", backdropPath: "", overview: "Overview 3", releaseDate: "2019-06-03", title: "Movie 3", genres: ["genre 3"])
+
+]
